dynamic_tables/requirements.txt
Django
djangorestframework
psycopg2
gunicorn
whitenoise
